Ensured Turn PID and FollowPath are always in high gear
In theory, TurnPID would be better to be in low gear to get more power,
but the time it takes to shift gears messes up the algorithm.
Maintaining the same gear in both FollowPath and TurnPID removes this
error, and since we want FollowPath to be faster than a snail, we pick
high gear.