Added rumble when the driver tries to intake but the elevator is up
Whenever the elevator is up above 14", we don't want to intake any cubes
since that would get the cubes stuck below the elevator and make a mess.
This behaviour already existed. However, the drivers sometimes could not
tell that the elevator was not at the bottom, and therefore were
confused why the robot wasn't intaking. Adding rumble feedback when this
occurs tells the drivers they are doing something wrong.