Fixed elevator climb mode.
When the elevator is in low gear (Climb mode), disable the PID and
control the elevator with open loop control. Our current PID tunings
have a very small lower bound of output power (-20%) since for normal
operation, we don't need much power to move the elevator down, we let
gravity help. This means the PID tunings are very inadequate for
climbing, though, so the simple fix is to just do climbing in open loop
control